Fixed object pooling issues!
object_pool class:
Now uses a List of gameObjects rather than an array. This ensures new objects can be added to the pool dynamically if needed without having to re-allocate all the memory.
The object_pool List is now also public, as it needs to be accessed via other classes.

plane_fire class:
Please read comments in this code!
Now has a public Transform which can be attached to specify the starting position of any fired bullets.
Rewrote the firing code. Now properly searches for an available bullet in the pool, fires said bullet, and deactivates it after a certain time period has elapsed.
Despawn timer Coroutine now takes a GameObject as a parameter, so it knows which object to deactivate after the timer is up.